nbs_analyzer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nbs-analyzer"
version = "1.0.0"
description = "Deterministic NbS Analyzer for Storyline A (Risk-First) and Storyline B (Benefits/Equity-First) analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CATIE Consultancy", email = "consultancy@example.com" }
]
keywords = ["nbs", "nature-based-solutions", "climate", "analysis", "risk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.7.0",
    "jinja2>=3.1.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
nbs-analyze = "nbs_analyzer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/nbs_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
